+++ tests/_Integration/Server.php
@@ -72,3 +72,19 @@ if (str_starts_with($route, '/too-many-requests')) {
 
     return include(__DIR__ . '/_Server/TooManyRequests.php');
 }
+
+if (str_starts_with($route, '/client-error-response')) {
+    $responseCodes = [400, 401, 404, 405, 410];
+
+    http_response_code($responseCodes[rand(0, 4)]);
+
+    return;
+}
+
+if (str_starts_with($route, '/server-error-response')) {
+    $responseCodes = [500, 502, 503, 505, 521];
+
+    http_response_code($responseCodes[rand(0, 4)]);
+
+    return;
+}